Add test cases for "AppName/2.0 (iPad; iOS 8.0; Scale/2.00)"
I've been seeing this user agent string in Sentry and the errors logged
with it do not have any browser or OS tags. From what I can tell of
Sentry, they are using the ua-parser project for user agent string
parsing.

Before digging in too deep on this issue, I decided to start by adding
some test cases for this user agent string to verify whether it is
working as expected or not. This will tell me if I should look here or
elsewhere.